+++ owncloudApp/src/main/java/com/owncloud/android/presentation/spaces/SpacesListFragment.kt
@@ -50,6 +50,7 @@ import com.owncloud.android.extensions.toTitleStringRes
 import com.owncloud.android.presentation.capabilities.CapabilityViewModel
 import com.owncloud.android.utils.DisplayUtils
 import com.owncloud.android.presentation.common.UIResult
+import com.owncloud.android.presentation.spaces.createspace.CreateSpaceDialogFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.koin.core.parameter.parametersOf
 import timber.log.Timber
@@ -103,6 +104,12 @@ class SpacesListFragment : SpacesListAdapter.SpacesListAdapterListener, Fragment
         binding.swipeRefreshSpacesList.setOnRefreshListener {
             spacesListViewModel.refreshSpacesFromServer()
         }
+
+        binding.fabCreateSpace.setOnClickListener {
+            val dialog = CreateSpaceDialogFragment.newInstance(requireArguments().getString(BUNDLE_ACCOUNT_NAME))
+            dialog.show(requireActivity().supportFragmentManager, DIALOG_CREATE_SPACE)
+            binding.fabCreateSpace.isFocusable = false
+        }
         setTextHintRootToolbar()
     }
 
@@ -136,7 +143,8 @@ class SpacesListFragment : SpacesListAdapter.SpacesListAdapterListener, Fragment
                 val accountName = requireArguments().getString(BUNDLE_ACCOUNT_NAME)
                 when (val uiResult = event.peekContent()) {
                     is UIResult.Success -> {
-                        Timber.d ("The account id for $accountName is: ${uiResult.data}")
+                        Timber.d("The account id for $accountName is: ${uiResult.data}")
+                        uiResult.data?.let { spacesListViewModel.getUserPermissions(it) }
                     }
                     is UIResult.Loading -> { }
                     is UIResult.Error -> {
@@ -146,6 +154,22 @@ class SpacesListFragment : SpacesListAdapter.SpacesListAdapterListener, Fragment
             }
         }
 
+        collectLatestLifecycleFlow(spacesListViewModel.userPermissions) { event ->
+            event?.let {
+                val accountName = requireArguments().getString(BUNDLE_ACCOUNT_NAME)
+                when (val uiResult = event.peekContent()) {
+                    is UIResult.Success -> {
+                        Timber.d("The permissions for $accountName are: ${uiResult.data}")
+                        uiResult.data?.let { binding.fabCreateSpace.isVisible = it.contains(DRIVES_CREATE_ALL_PERMISSION) }
+                    }
+                    is UIResult.Loading -> { }
+                    is UIResult.Error -> {
+                        Timber.e(uiResult.error, "Failed to retrieve user permissions for account $accountName")
+                        binding.fabCreateSpace.isVisible = false
+                    }
+                }
+            }
+        }
     }
 
     private fun showOrHideEmptyView(spacesList: List<OCSpace>) {
@@ -236,6 +260,10 @@ class SpacesListFragment : SpacesListAdapter.SpacesListAdapterListener, Fragment
         const val BUNDLE_KEY_CLICK_SPACE = "BUNDLE_KEY_CLICK_SPACE"
         const val BUNDLE_SHOW_PERSONAL_SPACE = "showPersonalSpace"
         const val BUNDLE_ACCOUNT_NAME = "accountName"
+        const val DRIVES_CREATE_ALL_PERMISSION = "Drives.Create.all"
+
+        private const val DIALOG_CREATE_SPACE = "DIALOG_CREATE_SPACE"
+
         fun newInstance(
             showPersonalSpace: Boolean,
             accountName: String