+++ owncloudComLibrary/src/main/java/com/owncloud/android/lib/resources/spaces/services/OCSpacesService.kt
@@ -24,10 +24,15 @@ package com.owncloud.android.lib.resources.spaces.services
 
 import com.owncloud.android.lib.common.OwnCloudClient
 import com.owncloud.android.lib.common.operations.RemoteOperationResult
+import com.owncloud.android.lib.resources.spaces.CreateRemoteSpaceOperation
 import com.owncloud.android.lib.resources.spaces.GetRemoteSpacesOperation
 import com.owncloud.android.lib.resources.spaces.responses.SpaceResponse
 
 class OCSpacesService(override val client: OwnCloudClient) : SpacesService {
     override fun getSpaces(): RemoteOperationResult<List<SpaceResponse>> =
         GetRemoteSpacesOperation().execute(client)
+
+    override fun createSpace(spaceName: String, spaceSubtitle: String, spaceQuota: Long): RemoteOperationResult<SpaceResponse> =
+        CreateRemoteSpaceOperation(spaceName, spaceSubtitle, spaceQuota).execute(client)
+
 }