+++ owncloudComLibrary/src/main/java/com/owncloud/android/lib/resources/users/services/implementation/OCUserService.kt
@@ -31,6 +31,7 @@ import com.owncloud.android.lib.common.operations.RemoteOperationResult
 import com.owncloud.android.lib.resources.users.GetRemoteUserAvatarOperation
 import com.owncloud.android.lib.resources.users.GetRemoteUserIdOperation
 import com.owncloud.android.lib.resources.users.GetRemoteUserInfoOperation
+import com.owncloud.android.lib.resources.users.GetRemoteUserPermissionsOperation
 import com.owncloud.android.lib.resources.users.GetRemoteUserQuotaOperation
 import com.owncloud.android.lib.resources.users.RemoteAvatarData
 import com.owncloud.android.lib.resources.users.RemoteUserInfo
@@ -48,5 +49,8 @@ class OCUserService(override val client: OwnCloudClient) : UserService {
 
     override fun getUserId(): RemoteOperationResult<String> =
         GetRemoteUserIdOperation().execute(client)
+
+    override fun getUserPermissions(accountId: String): RemoteOperationResult<List<String>> =
+        GetRemoteUserPermissionsOperation(accountId).execute(client)
 
 }