+++ owncloudData/src/main/java/com/owncloud/android/data/spaces/repository/OCSpacesRepository.kt
@@ -5,7 +5,7 @@
  * @author Juan Carlos Garrote Gascón
  * @author Jorge Aguado Recio
  *
- * Copyright (C) 2024 ownCloud GmbH.
+ * Copyright (C) 2025 ownCloud GmbH.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -76,5 +76,9 @@ class OCSpacesRepository(
 
     override fun getWebDavUrlForSpace(accountName: String, spaceId: String?): String? =
         localSpacesDataSource.getWebDavUrlForSpace(accountName = accountName, spaceId = spaceId)
+
+    override fun createSpace(accountName: String, spaceName: String, spaceSubtitle: String, spaceQuota: Long) {
+        remoteSpacesDataSource.createSpace(accountName, spaceName, spaceSubtitle, spaceQuota)
+    }
 
 }