+++ owncloudData/src/main/java/com/owncloud/android/data/user/datasources/RemoteUserDataSource.kt
@@ -30,4 +30,5 @@ interface RemoteUserDataSource {
     fun getUserQuota(accountName: String): UserQuota
     fun getUserAvatar(accountName: String): UserAvatar
     fun getUserId(accountName: String): String
+    fun getUserPermissions(accountName: String, accountId: String): List<String>
 }