+++ owncloudDomain/src/main/java/com/owncloud/android/domain/spaces/SpacesRepository.kt
@@ -3,8 +3,9 @@
  *
  * @author Abel García de Prada
  * @author Juan Carlos Garrote Gascón
+ * @author Jorge Aguado Recio
  *
- * Copyright (C) 2023 ownCloud GmbH.
+ * Copyright (C) 2025 ownCloud GmbH.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -33,4 +34,5 @@ interface SpacesRepository {
     fun getSpaceWithSpecialsByIdForAccount(spaceId: String?, accountName: String): OCSpace
     fun getSpaceByIdForAccount(spaceId: String?, accountName: String): OCSpace?
     fun getWebDavUrlForSpace(accountName: String, spaceId: String?): String?
+    fun createSpace(accountName: String, spaceName: String, spaceSubtitle: String, spaceQuota: Long)
 }